requirements.txt
Flask
google-cloud-bigquery
google-cloud-logging
pandas
slack-sdk
uvicorn
fastapi
db-dtypes